requirements.txt
discord.py>=2.3.0
mcp>=0.1.0
google-cloud-vision
google-generativeai
aiohttp